Added tests for simple auction and remove unused test files
